Accuracy: minor changes

+++ accuracy_checker/include/meshac/type_definition.hpp
@@ -27,9 +27,9 @@ namespace meshac {
 
         bool is(int a, int b, int c)
         {
-            return (a == vs[0] && b == vs[1] && c == vs[2]) ||
-                (b == vs[0] && c == vs[1] && a == vs[2]) ||
-                (c == vs[0] && a == vs[1] && b == vs[2]);
+            return (a == (int)vs[0] && b == (int)vs[1] && c == (int)vs[2]) ||
+                (b == (int)vs[0] && c == (int)vs[1] && a == (int)vs[2]) ||
+                (c == (int)vs[0] && a == (int)vs[1] && b == (int)vs[2]);
         }
 
     } FaceIndex;
